requirements.txt
asgiref==3.5.2
attrs==22.2.0
beautifulsoup4==4.11.2
black==23.7.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
Django==3.2.15
django-bootstrap5==22.2
django_debug_toolbar==3.8.1
Faker==12.0.1
flake8==4.0.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.6.1
mixer==7.2.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
Pillow==9.3.0
platformdirs==3.9.1
pluggy==1.0.0
prettytable==3.7.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pytest==7.1.3
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
pytest-subtests==0.9.0
python-dateutil==2.8.2
pytils==0.4.1
pytz==2022.7
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
yapf==0.32.0
